+++ perl/FSM/Scheduler/ISF/LoweringIR.pm
@@ -6261,9 +6261,17 @@ sub _validate_repeat_body_spawn_subset {
                 && $pending_local_do_before_drain
                 && !$pending_generated_do_before_drain
                 && @pending_spawns;
+            my $allowed_branch_spawn_after_prior_await_any_generated_child_do_before_drain =
+                ($when_body_repeat || $switch_branch_repeat)
+                && $awaiting_multi_pending_drain
+                && !$pending_local_do_before_drain
+                && $pending_generated_do_before_drain
+                && (($pending_generated_do_kind_before_drain // '') eq 'generated-child do')
+                && @pending_spawns;
             confess "Transaction '$tn': repeat-body spawn cannot follow multi-pending await_any before same-body await_all drains outstanding children\n"
                 if $awaiting_multi_pending_drain
-                    && !$allowed_branch_spawn_after_prior_await_any_local_do_before_drain;
+                    && !$allowed_branch_spawn_after_prior_await_any_local_do_before_drain
+                    && !$allowed_branch_spawn_after_prior_await_any_generated_child_do_before_drain;
             my $allowed_branch_spawn_after_local_do_before_drain =
                 ($when_body_repeat || $switch_branch_repeat)
                 && $pending_local_do_before_drain